Refactor collapsed chain utilities
