+++ modules/graph-layers/src/core/interaction-manager.ts
@@ -7,74 +7,13 @@ import {Edge} from '../graph/edge';
 import {Node} from '../graph/node';
 import {GraphEngine} from './graph-engine';
 import {log} from '../utils/log';
+import {
+  resolveChainInteractionSource,
+  type ChainInteractionSource
+} from '../utils/collapsed-chains';
 
-export type ChainInteractionSource =
-  | 'node'
-  | 'collapsed-marker'
-  | 'expanded-marker'
-  | 'collapsed-outline'
-  | 'expanded-outline';
-
-function resolveLayerId(layer: any): string {
-  if (!layer) {
-    return '';
-  }
-  if (typeof layer.id === 'string') {
-    return layer.id;
-  }
-  if (typeof layer.props?.id === 'string') {
-    return layer.props.id;
-  }
-  return '';
-}
-
-function classifyLayer(layer: any): ChainInteractionSource | null {
-  let current = layer ?? null;
-  while (current) {
-    const layerId = resolveLayerId(current);
-    if (layerId.includes('collapsed-chain-markers')) {
-      return 'collapsed-marker';
-    }
-    if (layerId.includes('expanded-chain-markers')) {
-      return 'expanded-marker';
-    }
-    if (layerId.includes('collapsed-chain-outlines')) {
-      return 'collapsed-outline';
-    }
-    if (layerId.includes('expanded-chain-outlines')) {
-      return 'expanded-outline';
-    }
-    current = current.parent ?? null;
-  }
-  return null;
-}
-
-export function resolveChainInteractionSource(info: any): ChainInteractionSource {
-  if (!info) {
-    return 'node';
-  }
-
-  const layersToCheck = [] as any[];
-  if (info.layer || info.sourceLayer) {
-    if (info.layer) {
-      layersToCheck.push(info.layer);
-    }
-    if (info.sourceLayer && info.sourceLayer !== info.layer) {
-      layersToCheck.push(info.sourceLayer);
-    }
-  } else {
-    layersToCheck.push(info);
-  }
-
-  for (const layer of layersToCheck) {
-    const classification = classifyLayer(layer);
-    if (classification) {
-      return classification;
-    }
-  }
-
-  return 'node';
-}
+export {resolveChainInteractionSource};
+export type {ChainInteractionSource};
 
 export function shouldToggleCollapsedChain(
   isCollapsed: boolean,
@@ -201,15 +140,16 @@ export class InteractionManager {
           // eslint-disable-next-line max-depth
           if (shouldToggleCollapsedChain(isCollapsed, interactionSource)) {
             const action = isCollapsed ? 'expand' : 'collapse';
+            const chainIdStr = String(chainId);
             log.log(
               0,
-              `InteractionManager: ${action} chain ${chainId} via ${interactionSource}`
+              `InteractionManager: ${action} chain ${chainIdStr} via ${interactionSource}`
             );
             // eslint-disable-next-line no-console
             console.log(
-              `InteractionManager: ${action} chain ${chainId} via ${interactionSource}`
+              `InteractionManager: ${action} chain ${chainIdStr} via ${interactionSource}`
             );
-            layout.toggleCollapsedChain(String(chainId));
+            layout.toggleCollapsedChain(chainIdStr);
             this._lastInteraction = Date.now();
             this.notifyCallback();
             // eslint-disable-next-line max-depth
